cubecl_reduce

Module instructions

Modules

fuse_accumulator_inplace

reduce_inplace

reduce_shared_inplace

Structs

ArgAccumulator

A pair of shared memory used for [ArgMax] and [ArgMin].

ArgAccumulatorExpand

ArgMax

Compute the coordinate of the maximum item returning the smallest coordinate in case of equality.

ArgMin

Compute the coordinate of the maximum item returning the smallest coordinate in case of equality.

Mean

Prod

Sum

Traits

Reduce

ReduceInstruction

An instruction for a reduce algorithm that works with Line.

SharedAccumulator

A simple trait that abstract over a single or multiple shared memory.

Functions

fuse_accumulator_inplace

reduce_inplace

reduce_shared_inplace