cubecl_reduce::instructions

Trait ReduceInstruction

pub trait ReduceInstruction<In: Numeric>:
    Send
    + Sync
    + 'static
    + Debug {
    type AccumulatorItem: CubeType;
    type SharedAccumulator: SharedAccumulator<In, Item = Self::AccumulatorItem>;

    // Required methods
    fn null_input(line_size: u32) -> Line<In>;
    fn null_accumulator(line_size: u32) -> Self::AccumulatorItem;
    fn assign_accumulator(
        destination: &mut Self::AccumulatorItem,
        source: &Self::AccumulatorItem,
    );
    fn reduce(
        accumulator: &Self::AccumulatorItem,
        item: Line<In>,
        coordinate: Line<u32>,
        use_planes: bool,
    ) -> Self::AccumulatorItem;
    fn fuse_accumulators(
        lhs: Self::AccumulatorItem,
        rhs: Self::AccumulatorItem,
    ) -> Self::AccumulatorItem;
    fn merge_line<Out: Numeric>(
        accumulator: Self::AccumulatorItem,
        shape_axis_reduce: u32,
    ) -> Out;
    fn to_output_perpendicular<Out: Numeric>(
        accumulator: Self::AccumulatorItem,
        shape_axis_reduce: u32,
    ) -> Line<Out>;
    fn __expand_null_input(
        context: &mut CubeContext,
        line_size: u32,
    ) -> <Line<In> as CubeType>::ExpandType;
    fn __expand_null_accumulator(
        context: &mut CubeContext,
        line_size: u32,
    ) -> <Self::AccumulatorItem as CubeType>::ExpandType;
    fn __expand_assign_accumulator(
        context: &mut CubeContext,
        destination: <Self::AccumulatorItem as CubeType>::ExpandType,
        source: <Self::AccumulatorItem as CubeType>::ExpandType,
    ) -> <() as CubeType>::ExpandType;
    fn __expand_reduce(
        context: &mut CubeContext,
        accumulator: <Self::AccumulatorItem as CubeType>::ExpandType,
        item: <Line<In> as CubeType>::ExpandType,
        coordinate: <Line<u32> as CubeType>::ExpandType,
        use_planes: bool,
    ) -> <Self::AccumulatorItem as CubeType>::ExpandType;
    fn __expand_fuse_accumulators(
        context: &mut CubeContext,
        lhs: <Self::AccumulatorItem as CubeType>::ExpandType,
        rhs: <Self::AccumulatorItem as CubeType>::ExpandType,
    ) -> <Self::AccumulatorItem as CubeType>::ExpandType;
    fn __expand_merge_line<Out: Numeric>(
        context: &mut CubeContext,
        accumulator: <Self::AccumulatorItem as CubeType>::ExpandType,
        shape_axis_reduce: <u32 as CubeType>::ExpandType,
    ) -> <Out as CubeType>::ExpandType;
    fn __expand_to_output_perpendicular<Out: Numeric>(
        context: &mut CubeContext,
        accumulator: <Self::AccumulatorItem as CubeType>::ExpandType,
        shape_axis_reduce: <u32 as CubeType>::ExpandType,
    ) -> <Line<Out> as CubeType>::ExpandType;

}

An instruction for a reduce algorithm that works with Line.

See a provided implementation, such as [Sum] or [ArgMax] for an example how to implement this trait for a custom instruction.

A reduction works at three levels. First, it takes input data of type In and reduce them with their coordinate into an AccumulatorItem. Then, multiple AccumulatorItem are possibly fused together into a single accumulator that is converted to the expected output type.

Required Associated Types

type AccumulatorItem: CubeType

The intermediate state into which we accumulate new input elements. This is most likely a Line<T> or a struct or tuple of lines.

type SharedAccumulator: SharedAccumulator<In, Item = Self::AccumulatorItem>

When multiple agents are collaborating to reduce a single slice, we need a share accumulator to store multiple AccumulatorItem. This is most likely a SharedMemory<Line<T>> or a struct or tuple of lined shared memories.

Required Methods

fn null_input(line_size: u32) -> Line<In>

A input such that Self::reduce(accumulator, Self::null_input(), coordinate, use_planes) is guaranteed to return accumulator unchanged for any choice of coordinate.

fn null_accumulator(line_size: u32) -> Self::AccumulatorItem

A accumulator such that Self::fuse_accumulators(accumulator, Self::null_accumulator() always returns is guaranteed to return accumulator unchanged.

fn assign_accumulator( destination: &mut Self::AccumulatorItem, source: &Self::AccumulatorItem, )

Assign the value of source into destination. In spirit, this is equivalent to destination = source;, but this syntax is not currently supported by CubeCL.

fn reduce( accumulator: &Self::AccumulatorItem, item: Line<In>, coordinate: Line<u32>, use_planes: bool, ) -> Self::AccumulatorItem

If use_planes is true, reduce all the item and coordinate within the accumulator. Else, reduce the given item and coordinate into the accumulator.

fn fuse_accumulators( lhs: Self::AccumulatorItem, rhs: Self::AccumulatorItem, ) -> Self::AccumulatorItem

Reduce two accumulators into a single accumulator.

fn merge_line<Out: Numeric>( accumulator: Self::AccumulatorItem, shape_axis_reduce: u32, ) -> Out

Reduce all elements of the accumulator into a single output element of type Out.

fn to_output_perpendicular<Out: Numeric>( accumulator: Self::AccumulatorItem, shape_axis_reduce: u32, ) -> Line<Out>

Convert each element of the accumulator into the expected output element of type Out.

fn __expand_null_input( context: &mut CubeContext, line_size: u32, ) -> <Line<In> as CubeType>::ExpandType

fn __expand_null_accumulator( context: &mut CubeContext, line_size: u32, ) -> <Self::AccumulatorItem as CubeType>::ExpandType

fn __expand_assign_accumulator( context: &mut CubeContext, destination: <Self::AccumulatorItem as CubeType>::ExpandType, source: <Self::AccumulatorItem as CubeType>::ExpandType, ) -> <() as CubeType>::ExpandType

fn __expand_reduce( context: &mut CubeContext, accumulator: <Self::AccumulatorItem as CubeType>::ExpandType, item: <Line<In> as CubeType>::ExpandType, coordinate: <Line<u32> as CubeType>::ExpandType, use_planes: bool, ) -> <Self::AccumulatorItem as CubeType>::ExpandType

fn __expand_fuse_accumulators( context: &mut CubeContext, lhs: <Self::AccumulatorItem as CubeType>::ExpandType, rhs: <Self::AccumulatorItem as CubeType>::ExpandType, ) -> <Self::AccumulatorItem as CubeType>::ExpandType

fn __expand_merge_line<Out: Numeric>( context: &mut CubeContext, accumulator: <Self::AccumulatorItem as CubeType>::ExpandType, shape_axis_reduce: <u32 as CubeType>::ExpandType, ) -> <Out as CubeType>::ExpandType

fn __expand_to_output_perpendicular<Out: Numeric>( context: &mut CubeContext, accumulator: <Self::AccumulatorItem as CubeType>::ExpandType, shape_axis_reduce: <u32 as CubeType>::ExpandType, ) -> <Line<Out> as CubeType>::ExpandType

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.