cubecl_reduce::primitives

Function reduce_slice

pub fn reduce_slice<N: Numeric, R: ReduceInstruction<N>>(
    items: Slice<Line<N>>,
    range: ReduceRange,
    line_size: u32,
    line_mode: LineMode,
) -> R::AccumulatorItem

Use an individual unit to reduce the items with the specified range. That is, this will reduces items[range.start], items[range.start + range.step] until items[range.end] (exclusive).

This reduces using the given line_mode but doesn’t reduce the accumulator itself.

Since each individual unit performs a reduction, this function is meant to be called with either a different items for each unit, a different range or both based on ABSOLUTE_UNIT_POS.